split title in sidebar and fix click event
