+++ src/components/Layout/Sidebar_Layout.js
@@ -1,55 +1,12 @@
 import React, { useState } from "react";
-import { Link } from "react-router-dom";
 
 import Homebutton from "./Homebutton_Layout";
 import Dashboardbutton from "./Dashboardbutton_Layout";
 import Aboutbutton from "./Aboutbutton_Layout";
-
-import icon from "../../media/icon.png";
-import icon_dark from "../../media/icon_dark.png";
-
-const Title = React.memo((props) => {
-  //console.log("Title: " + props);
-
-  if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
-    return (
-      <li
-        className="inline-flex items-center p-2 mr-4 
-        text-gray-700 dark:text-white"
-      >
-        <img
-          src={icon_dark}
-          className="relative rounded-lg p-4 h-20 w-20"
-          alt=""
-          loading="lazy"
-        ></img>
-        <Link to="/" className="text-xl font-bold uppercase tracking-wide">
-          Asian Food in Berlin
-        </Link>
-      </li>
-    );
-  } else {
-    return (
-      <li
-        className="inline-flex items-center p-2 mr-4 
-        text-gray-700 dark:text-white"
-      >
-        <img
-          src={icon}
-          className="relative rounded-lg p-4 h-20 w-20"
-          alt=""
-          loading="lazy"
-        ></img>
-        <Link to="/" className="text-xl font-bold uppercase tracking-wide">
-          Asian Food in Berlin
-        </Link>
-      </li>
-    );
-  }
-});
+import Title from "./Title_Layout";
 
 const Sidebar = (props) => {
-  console.log("Sidebar: " + props);
+  //console.log("Sidebar: " + props);
 
   const change_sidebar_state = props.change_sidebar_state.bind(this);
   let sidebar_state = props.sidebar_state;
@@ -84,7 +41,9 @@ const Sidebar = (props) => {
       overflow-hidden
       `}
     >
-      <Title></Title>
+      <Title
+        change_sidebar_item_selected={change_sidebar_item_selected}
+      ></Title>
       <nav className="mt-10">
         <Homebutton
           siderbar_item_selected={siderbar_item_selected}
